whatsapp-bot/requirements.txt
fastapi>=0.110
uvicorn>=0.29
twilio>=9.0
google-genai>=1.0
google-api-python-client==2.134.0
google-auth>=2.0
httpx>=0.27
python-dotenv==1.0.1
pydantic>=2.0
pydantic-settings>=2.0
pytz==2024.1
pytest
pytest-asyncio
